Move `GetGitRepositoryPath`
