+++ src/Files.App/Utils/Git/GitHelpers.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Files Community
+// Copyright (c) Files Community
 // Licensed under the MIT License.
 
 using Files.App.Dialogs;
@@ -13,6 +13,14 @@ namespace Files.App.Utils.Git
 {
 	internal static partial class GitHelpers
 	{
+		// The implementation of the version control interface; it's hardcoded right now but will be made configurable in the future (#16738)
+		private static readonly LibGit2 _implementation = new LibGit2(); // TODO: Replace with IVersionControl abstraction when it is complete
+
+		/// <inheritdoc cref="IVersionControl.GetGitRepositoryPath(string?,string)"/>
+		public static string? GetGitRepositoryPath(string? path, string root) => _implementation.GetGitRepositoryPath(path, root);
+
+		#region Legacy implementation
+
 		private static readonly StatusCenterViewModel StatusCenterViewModel = Ioc.Default.GetRequiredService<StatusCenterViewModel>();
 
 		private const string GIT_RESOURCE_NAME = "Files:https://github.com";
@@ -59,43 +67,6 @@ namespace Files.App.Utils.Git
 		public static event PropertyChangedEventHandler? IsExecutingGitActionChanged;
 
 		public static event EventHandler? GitFetchCompleted;
-
-		public static string? GetGitRepositoryPath(string? path, string root)
-		{
-			if (string.IsNullOrEmpty(root))
-				return null;
-
-			if (root.EndsWith('\\'))
-				root = root.Substring(0, root.Length - 1);
-
-			if (string.IsNullOrWhiteSpace(path) ||
-				path.Equals(root, StringComparison.OrdinalIgnoreCase) ||
-				path.Equals("Home", StringComparison.OrdinalIgnoreCase) ||
-				ShellStorageFolder.IsShellPath(path))
-			{
-				return null;
-			}
-
-			try
-			{
-				if (IsRepoValid(path))
-					return path;
-				else
-				{
-					var parentDir = PathNormalization.GetParentDir(path);
-					if (parentDir == path)
-						return null;
-					else
-						return GetGitRepositoryPath(parentDir, root);
-				}
-			}
-			catch (Exception ex) when (ex is LibGit2SharpException or EncoderFallbackException)
-			{
-				_logger.LogWarning(ex.Message);
-
-				return null;
-			}
-		}
 
 		public static string GetOriginRepositoryName(string? path)
 		{
@@ -768,6 +739,7 @@ namespace Files.App.Utils.Git
 			}
 		}
 
+		// Method already moved into abstraction
 		private static bool IsRepoValid(string path)
 		{
 			return SafetyExtensions.IgnoreExceptions(() => Repository.IsValid(path));
@@ -967,5 +939,7 @@ namespace Files.App.Utils.Git
 
 		[GeneratedRegex(@"^(?:https?:\/\/)?(?:www\.)?(?<domain>github|gitlab)\.com\/(?<user>[^\/]+)\/(?<repo>[^\/]+?)(?=\.git|\/|$)(?:\.git)?(?:\/)?", RegexOptions.IgnoreCase)]
 		private static partial Regex GitHubRepositoryRegex();
+
+		#endregion
 	}
 }
